backend/requirements.txt
flask
flask-cors
python-dotenv
supabase
python-dateutil
web3
